all buttons fully functional

+++ backend/app/models/evaluation.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, ForeignKey, UniqueConstraint
+from sqlalchemy import Column, Integer, Float, ForeignKey, UniqueConstraint
 from app.core.database import Base
 
 class Evaluation(Base):
@@ -7,7 +7,7 @@ class Evaluation(Base):
     id = Column(Integer, primary_key=True)
     judge_id = Column(Integer, ForeignKey("judges.id"), nullable=False)
     team_id = Column(Integer, ForeignKey("teams.id"), nullable=False)
-    score = Column(Integer, nullable=False)  # 1–10
+    score = Column(Float, nullable=False)  # 1–10 (Allows decimals like 8.7)
 
     __table_args__ = (
         UniqueConstraint("judge_id", "team_id", name="unique_judge_team_eval"),
